(PUP-10223) Allow node definitions in apply blocks
Previously, we explicitly forbade node definitions in plans entirely as
Bolt didn't know how to interpret them. This changes to allowing node
definitions in apply blocks so that they can be interpreted by Bolt.

When validating node definitions, we now check whether we're in an apply
block and fail if we aren't. Otherwise, we defer to the standard
validator. That validation will fail however, because definitions are
only allowed at the top level. For the intents and purposes of a plan,
an apply block is effectively a top-level construct, as it will be used
as the top-level input to a catalog compilation. Therefore, we now allow
node definitions to be defined directly within apply blocks (but not
nested any deeper than that).

The last piece of the puzzle is to change the plan instantiator to not
treat these node definitions nested within apply blocks as though they
are top-level definitions that would constitute a single file illegaly
making multiple definitions. We now simply remove node definitions from
consideration when checking whether the plan file has multiple
definitions.