(maint) Update tests to actually not require a master host when running
Prior to this commit, some of the tests had code that required a master
host before we could specify the tags. The tags are important to define
first because they let us know when the test requires a provisioned
master. There are scenarios where we do not bother to provision a
puppetserver (aka master host). We need to be able to skip all tests
that require a server. These tests are generally tagged with 'server'.
However, if there is code evaluated before the tags are defined that
reference the master host, that test will fail. This commit fixes that.

This commit also adds 'server' tags if they're missing from a test.