(maint) remove references to master where we don't need them
As this test doesn't require a master host to exist, we should remove
all references to master. We have the option to run tests without even
bothering to provision a master. When we do this, any reference to a
master in the tests will cause that test to fail. We want to be able to
run this test without provisioning a master host, so we need to remove
the reference to a master.