(PUP-8900) Use HTTP client to upload facts
Previously, `puppet facts upload` only tried the first entry in `server_list`,
but never fell back to other entries in the list.

Now it uses the http client which attempts to connect to each entry in the
`server_list`, so that `puppet facts upload` behaves the same as `puppet
agent` for SRV, `server_list` and `server` resolution.

There's one minor change to the log message to remove the extra ':'

Faces will output to the console whatever the `when_invoked` block
returns, so return an explicit `nil` which is what `indirection.save`
did.