(maint) [WIP] Use canonical forge_name
This just corrects the separator on the `forge_name` attribute of
modules now that we've standardized on the `-` rather than the `/`. All
the tests having to do with modules pass. I don't know if I'm
misunderstanding the full scope of what this method is intended for
though, so please don't merge without review by someone who knows that
part of the code and the history of the forge name inconsistencies.