(PUP-10240) Add development gem dependencies to gemspec
Move development gem dependencies from the Gemfile to .gemspec. As a result, the
dependencies are included in the published puppet gem.

It is no longer possible to specify `RAKE_LOCATION` as an environment variable,
but that is no longer needed in 6.x, as we relaxed the constraint in 2c6bdea0f4.