(maint) Use memory terminus for facts indirection
Previously we were running facter many times during specs due to facter being
the default facts terminus.

This commit changes the default terminus to :memory when running tests, and
saves an empty facts node so it's available to tests. We only run facter once
in the facts integration test, and there we opt-in by setting the terminus
to `:facter`.

This commit only addresses tests that use the indirector to access facter. It
doesn't change how code or tests call Facter.value(key) or Facter[:key]
directly. More work is needed to ensure those calls don't cause facter to run.

It's not necessary to clear termini or caches before/after tests as that's
already done in the spec_helper.

Also delete useless test.

On OSX with 16 workers, rake parallel:spec now runs 25% faster:

    Before:
    Finished in 1 minute 36.44 seconds
    25020 examples, 0 failures, 51 pending

    After:
    Finished in 1 minute 11.3 seconds
    25019 examples, 0 failures, 51 pending

On Windows with 2 workers, it runs about 15% faster:

    Before:
    Finished in 10 minutes 6 seconds
    24765 examples, 0 failures, 104 pending

    After:
    Finished in 8 minutes 35 seconds
    24764 examples, 0 failures, 104 pending