(PUP-10255) Follow symlinks in environment path
This commit adds a `real_path` method that will follow symlinks in the given
directory, which is then used when getting the environment.conf and generating
metadata for a given environment. This enables us to have multiple code
dirs for a given environment, with the set `environmentpath` pointing to the
most recent version, which enables lockless deployments.