(PUP-10267) Eliminate need to pass mount and path separately
Previously, `mount_point` and `path` had to be passed separately and both were
required. But in practice we typically retrieve file metadata and content based
on an already "joined" path like '/modules/<module>/path'. Also the path can be
nil such as when recursively searching the '/plugins' mount for metadata.

To avoid every caller from having to split a path, only for the service to join
them back together, remove mount_point as a separate argument.