(PUP-10218) Puppet incorrectly detecting stale pidfile
When Puppet Agent is started as a LightWeight Process(LWP), and is
incorrectly terminated (eg. killed by KILL signal), decision for the
removal of stale pidfile lock (on POSIX) actually depends on the kernel,
operating system and sometimes even version due to the unreliability of
ps CLI arguments. The '-p' argument, used initially for all POSIX, is
indeed available for all (checked) versions of ps, but in some cases it
does not work as expected for LWP (printing nothing and returning error
code 1).

Depending on the kernel seemed to be the most clean approach to tackle
all the different scenarios of choosing the right ps argument. A
fallback was implemented for Linux operating systems where ps is
outdated and it does not have the -q (quick mode) parameter (such as Red
Hat 5) which seems to show the needed information even in case of LWP.