(PUP-10267) Remove extra HEAD request from vcr cassette
Previously, each VCR cassette had 6 request/responses, but one of the HEAD
requests was no longer valid, due to commit deac457fe3.

This commit removes the extra request/response. Note there are still 4
HEAD requests due to a bug in `HttpProxy.request_with_redirects` making
an unnecessary HEAD request prior to GET. And each HEAD request is
redirected in the VCR cassette, resulting in 4 total.

The request path for the GET request also changed due to a bug in
`request_with_redirects` which forces the GET request to use the path
from the redirected HEAD request (which ends in a slash).

Those issues will be cleaned up when we switch over to using the http
client to retrieve file content.