(PUP-10147) Require a block for Response#read_body
Previously, we were calling the streaming version of
`Puppet::HTTP::Response#read_body` and not supplying a block. This happens to
work (because that's how the underlying `Net::HTTP::Response#read_body` method
works), but that wasn't the intended contract for our response class.

This commit changes the client to use the non-streaming version in cases where
we know we're not streaming the response, such as PUT, DELETE, etc. It guards
against a missing block and adds tests for the response API.