(PUP-10142) Refactor settings default initialization
Previous to this commit, the puppet settings defaults were loaded
when the defaults.rb file was required. That design makes it
difficult to generate new settings objects; additionally, code
that executes by simply requiring the file seems less than ideal.

This change stops the code from executing by simply requiring the
defaults.rb file, and refactors it to become a class method that
modifies a settings object that is passed in. As a result of this
refactor, the `Puppet.define_settings` method is no longer required
as a bridge to the @@settings object and is set as deprecated.