(PUP-10260) Use http client to retrieve file content
The http client handles decompression, so delete the uncompress logic and call
to request_to_uri.

If a source content_uri is specified, then use its host/port when making the
static file content request.

If a source source_uri is specified, then use either the host/port or route to
the fileserver.

For http(s) URIs, call the lower-level `client.get` method, instead of
Puppet::Util::HttpProxy. As a result, we no longer make a second HEAD request
(which is redirected) prior to the GET. As a result, each VCR recording performs
3 requests:

   HEAD -> 301 redirect
   HEAD -> 200
   GET -> 200