(PUP-10260) Use http client for catalog requests
This preserves the existing behavior of the caller supplying serialized and URL
encoded facts as indirector request options, which the rest terminus needs
to decode and deserialize so it can pass a facts object to the http client.

As a result of this change, the agent will always POST for the catalog. In
theory, this could result in a behavior change, but only if the HTTP request
line is less than 1024 due to a ruby bug[1]. However, serialized and encoded
facts are sent as query parameters, and just core facts result in a request line
of over 8k. POST requests don't have this problem because the query parameters
are included in the request body.

It also preserves the behavior of setting/overwriting the catalog's name based
on what was requested. Some of the configurer tests rely on that behavior.

[1] https://bugs.ruby-lang.org/issues/3991