(PUP-10260) Use http client for catalog requests
This preserves the existing behavior of the caller supplying serialized and URL
encoded facts as indirector request options, which the rest terminus needs to
decode and deserialize so it can pass a facts object to the http client.

As a result of this change, the agent will always POST for the catalog. In
theory, it could result in a behavior change, but it would only happen if the
HTTP request line would have been less 1024. Since facts are sent as query
parameters and just core facts alone cause the request line to exceed 8k, puppet
always makes a POST request in practice.

It also preserves the behavior of setting/overwriting the catalog's name based
on what was requested. Some of the configurer tests rely on that behavior.

[1] https://bugs.ruby-lang.org/issues/3991