(PUP-10260) Use consistent node name
Previously, the test requested catalogs based on `node_name_value` (which
defaults to `certname`), but the returned catalog had a name `tester`. The tests
passed because the catalog rest terminus overrides the returned catalog with
whatever the caller requested.

That behavior was added in [1] but the rest terminus for 'certificate' already
does this, and it's not needed or desired for catalogs.

[1] https://github.com/puppetlabs/puppet/commit/089ac3e37dd1418751bc4dfe152e09fbacbc5122