(PUP-10260) Thread facts through catalog request
Return the collected facts object and thread that through to the places where we
need to request a new catalog and resolve deferred values. Previously we were
storing the facts object in the generic options hash (in `convert_with_facts`).

If the node request causes us to switch environments, then clear the facts, like
we do with query options, so that we resolve facts again after switching
the environment.