(PUP-10260) Pass facts for catalog in indirector request
Pass the facts object to the http client when requesting a catalog. The http
client should be responsible for serializing the request and deserializing the
response, instead of the configurer. Since the catalog request can be routed
to different termini, we can't remove the `facts` and `facts_format` from the
query parameters. So send both in the indirector request.