(PUP-10147) Collect per-site server versions
Process each response collecting per-site server versions. If server version
is >= 5 then JSON is supported. If version is >= 5.3.4, then the locales mount
is supported.

Because file metadata and content requests don't use the httpclient yet, we
have to lookup the server version in both the context and session.

The plugin integration test is unrealistic because it sets the metadata
and content termini to memory, so we can't determine what version the puppet
server is. So push a version onto the context during the test.