(PUP-10286) Delete the pool entry when the last connection is borrowed
Previously when the last connection for a site was borrowed, the pool still
contained a map entry for the site to an empty array of connections. The
entry persists until the pool is closed or a new connection is attempted
for the same site.

Now delete the map entry when the last connection for a site is borrowed.