(PUP-10294) Order dependent test failure with recent changes
A couple of tests from 'agent_spec.rb' were affecting one of the tests
from 'pidlock_spec.rb'. By mocking the Puppet::Util::Execution.execute
method in the newly introduced tests, to force returning what we were
expecting, no influence/order execution dependency is present anymore.
It was interesting to find out that the faulty version of the test
failed not only when run in this order/parallel execution, but also in
isolation, unless you ran something like:
>> bundle exec rspec /Users/luchi/puppet/spec/unit/pidlock_spec.rb

This is where by coincidence/pure luck, the path contains the word
'puppet'. This is important because it ends up in the arguments section
of the executed command. This combined with having the word 'ruby' in
the command name, was enough to pass the test. This happened due to the
following line of code from 'pidlock.rb':
'@lockfile.unlock unless procname =~ /ruby/ && args =~ /puppet/ ||
                         procname =~ /puppet(-.*)?$/'

In contrast, the test would've always failed if you ran it with a
relative path such as:
>> bundle exec rspec spec/unit/pidlock_spec.rb

Also, there was a case where the test could've passed even with a path
not containing the word 'puppet', by simply running beforehand:
>> bundle --path=vendor