(PUP-9566) Include user-specified headers in http requests
When we submit a request to the server, we now have the ability to add
in additional arbitrary headers. This should be isolated only to
requests made to a puppetserver, and these headers should not be
included in a request to an arbitrary external host.

These headers can be set with the `:http_extra_headers` setting in
Puppet. As illustrated in the various tests, this setting can be defined
in the following four forms:

  - string:
    `'header1:value,header2:value'`
  - array:
    `['header1:value', 'header2:value']`
  - array of arrays:
    `[['header1', 'value'], ['header2', 'value']]`
  - hash:
    `{'header1' => 'value', 'header2' => 'value'}`.

This commit also updates tests for the different services. It bundles
the test for the puppet profile header with the test for these custom
headers. These are both set in the same method, so a single test for
both makes sense. I also added the test for each different type of http
connection for the different services, to ensure consistent behavior
across all of them.