(maint) Use rspec around
Previously we were pushing server_agent_version for each rspec example, but
never popping explicitly, though it happens implicitly in the spec_helper when
we roll the context back to the previous mark.

Use rspec `around` so the context state is explicitly balanced. This only works
when the data being pushed does not rely on settings.