(PUP-10147) Don't resolve the compiler service for the locales mount
Previously, if the agent run failed to connect to the compiler service, then it
would exit without falling back to the cached catalog.

This occurred because we only store `server_agent_version` in the context for
successful REST requests. If the "plugins" and "pluginfacts" mounts fail to
connect, then `Session#route_to` triggered the resolution logic, and that would
raise after exhausting the various resolvers.

Now we check if the session supports the "locales" capability. If we haven't
connected to the compiler service successfully at some point earlier in the run,
then preserve the original 6.x behavior of skipping the "locales" mount.

Ideally, pluginsync would abort if any part of it failed, but that is a larger
issue, see PUP-1763.