(maint) Remove COMMITTERS.md
This file has not been edited, maintained or updated for a few years now and
contains a lot of out of date information that is covered in other places like
CONTRIBUTING.md, the code of conduct, the main repo readme, etc. so just
deleting it full stop to avoid folks landing on it and being confused by old
terms and branch names.