(PUP-10288) Ensure there is only one http pool
Previously, requests based on `Puppet::Network::HttpPool.http_instance` and the
http client used two different pools, so it was possible for connections to not
be reused when they could be.

This commit pushes the http client's pool onto the context for cases where we
explicitly want to use persistent connections.