(PUP-10288) Close the pool before sleeping
When running the ssl state machine, close the pool before sleeping so that we
don't keep connections open on the server. It also avoids trying to reuse a
connection that the remote side closed due to an idle timeout. This is similar
to the change made in 8639f0c5176b for PUP-10227.