(maint) Don't require types outside of the autoloader
Previously, puppet tests could generate warnings like

    warning: already initialized constant CREATORS

due to puppet types being loaded twice. This happened if a spec required the
type directly and it was later loaded by the autoloader. This is possible
because the autoloader does not consult $LOADED_FEATURES before checking to see
if a file should be loaded.

Updates tests to lookup a type on-demand using `Puppet::Type.type(:name)`.