(PUP-10314) Only pass a verifier for HTTPS connections
Previously, trying to start an HTTP (not HTTPS) connection at the debug log
level would fail trying to call `TCPSocket#ssl_version`. This is because the
client was always passing a verifier to the pool, but the pool was expecting a
nil verifier for HTTP connections.

Now use `Site#use_ssl?` to determine whether SSL mode is enabled or not. If so,
pass the verifier in as we were doing before, and only in that case do we setup
the connection and print the ssl debugging info.