(PUP-5069) Add include_system_store option
Adds an `include_system_store` argument to the http client, which is related to
the `ssl_context` argument.

If an `ssl_context` is specified and `include_system_context` are specified,
then raise, since they conflict.

If an `ssl_context` is specified, always use that.

If `include_system_context` is specified then create and memoize a system ssl
context, which trusts the puppet PKI and the system CA store, typically the
puppet-ca bundle component of puppet-agent. The context disables cert
revocation, since we don't have access to CRLs for arbitrary 3rd party HTTPS
servers and we don't support crlDistributionPoints extension. It also does not
use the client cert/private key.

If neither are specified, then only trust the puppet PKI, which is the default
behavior.