(PUP-10353) Replace keyword arguments to generic options hash
This commit removes keyword arguments in favor of a generic options
hash. Though keyword arguments are cleaner, we decided to go with the
options hash to give us more flexibility moving forward. As we add new
options, like report compression, we need to be able to do so without
changing the API. If the new option is there, we use it, otherwise, we
behave like we previously did. This makes everything a bit easier and
eliminates the need to have puppetserver enforce hard dependencies on
puppet versions because of API changes.