(PUP-10359) Resolve fileserver and report servers using server_list
The server_list resolver was committed before the fileserver and report servers
were. As a result, the server_list was not used to resolve them, causing the
reports/failover_master.rb acceptance test to fail when using an empty
puppet.conf, but an explicit `server_list` on the command line.

Use the server list resolver for all services, but exclude the ca and report
services if they are explicitly configured in puppet.conf or the command line.

The fileserver service is not excluded because there isn't a puppet setting
for it.