(PUP-10364) Add pp_owner ppRegCertExt OID
In Puppet-as-a-Service use cases it is common for individual nodes to be
owned by various different tenant, or customer, teams. None of the
registered ppRegCertExt OIDs we currently provide cleanly allow generic
specification of an owner in this context.

Provide a new OID, pp_owner, which generically fits this use case.