(WIP) Add Puppet::Transaction::Report#userdata
Without this patch the transaction report created by the agent does not
have an attribute to store user defined data.  This is a problem because
we need to store the following pieces of data in the report object to
make the data accessible via BQ and Data Studio:

 1. All classes
 2. Top level compiler variables from the catalog
 3. A way to identify if a change to code is included in the enforcement
    run.  Similar to config version, but may include additional data.

This patch addresses the problem by establishing an userdata attribute
which is a Hash map.  The schema of the value is not specified and left
to the user to define.  The value must be serializable to JSON encoding
so that the data may be loaded into BQ.

Next steps:

 1. Verify the attribute is present on the puppetserver.
 2. Verify a report processor has access to the data.
 3. Verify to_data_hash() includes userdata
 4. Add classes to the userdata.
 5. Add facts to the userdata.
 6. Add top level variables from the catalog to the userdata.