(PUP-10363) Cache the server list url not the service
The server_list resolver cached the service whose host and port it first
connected to, such as Puppet::HTTP::Service::Compiler. However, the next
request for a different service, would return the cached service, which wasn't
what the caller expected.

Instead cache the URL that we successfully connected to, and use that to
create the service of the correct type.