(PUP-10365) Include the system store for https file sources
Commit 11eee4e886d3a67b1973610727df17306607f164 changed puppet to use the http
client to retrieve file content from https sources. In the process, we were
using the default SSLContext, whose X509 store only trusts the puppet PKI, so
connections to 3rd party HTTPS servers failed.

Pass `include_system_store` so that we use the system SSLContext.

Add an integration test to verify an agent can download file content from an
HTTPS server whose CA cert is not the puppet CA, but is in the system CA
bundle. Also verify the download fails if the CA cert is not in the system CA
bundle.