(PUP-10367) Source files using full_path
This commit fixes a regression in 6.14.0 that prevented puppet apply from
sourcing a recursive directory.

`Metadata#path` is the absolute path from which metadata is collected. For files
and non-recursive directories, it is the same as `Metadata#full_path`.

However, for recursive directories, `path` and `full_path` are different for all
children of the directory. The `path` is always that of the base directory where
recursion started, while `full_path` is the absolute path of the subdirectory or
file within the base.