(maint) Add local el-repo fixture and fix rpm multiversion test
Since the upstream mirrors are not always reliable, create a local EL
repo to test the RPM multiple version functionality.

What I did was to download two `kernel-devel` RPMs, repackage them
with `rpmrebuild` to change their name and content, and put them in a
local repository.

This way we will not depend on external factors and our test should
always run.

The EL repo fixture can be extended and used in other tests as
well. The desired package has to be put inside `el-repo/RPMS`, and the
repository metadata needs to be regenerated with `createrepo`.