(PUP-10374) Download from the locales mount
puppet plugin download doesn't download from the locales mount, but
puppet agent does. This issue wasn't noticed due to the integration test
storing `server_agent_version` in the context, which doesn't happen when
using the http client.

This commit modifies the plugin handler to check to see if the session
supports the "locales" capability for either the "puppet" or "fileserver"
services. Also delete the useless face plugin_spec and replace with an
integration test that downloads from our test puppetserver.