(PUP-10235) Support enabling DNF modules
Add an optional `enable_only` parameter to the package type.

If set to `true` on a `dnfmodule` resource, it will cause the module
to be enabled instead of installed. The flag cannot be set together
with the `flavor` property, since they conflict.

DNF modules that have no default profile (and cannot be installed)
will be enabled by default, so there is no need to add `enable_only`.

To enable a module that has a default profile, `enabled_only` has to
be set.