(PUP-10397) Remove facts and vars from ScriptCompiler scope
This updates the scope for a `ScriptCompiler` to not contain any
`$facts`, `$trusted`, or `$server_facts`.

This work is to enable the use of the `lookup` function in Bolt plans
outside of apply blocks. Since the concept of `facts`, etc. do not make
sense in the scope of running a plan, these fields should not be
available. Removing these fields from the scope disables the use of
interpolations outside of apply blocks.

Related to puppetlabs/bolt#1403