(PUP-10379) Load Bolt project as a Puppet module
This adds a new class, Puppet::Node::BoltProject, which accepts an
absolute path to the Bolt project directory and loads it as a standalone
module (not a modulepath, but a single module). The class inherits from
Puppet::Node::Environment and only overrides the `modules` method to
insert the Bolt project "module" into the list of modules, and the
`initialize` method to accept the absolute path to the project. This is
only intended for use with Bolt, and is instantiated through the Pal
method `in_bolt_project`, which will either create a BoltProject class
if Bolt is configured to load the project dir, otherwise it will create
a Puppet::Node::Environment.