(PUP-10316) Update 'http' report processor to use the http client
Update the `http` report processor to post the report using the HTTP API. When
running `puppet apply`, use the default `Puppet::HTTP::Client` implementation.
When running in puppetserver, use the `ExternalClient`, which will delegate to
puppetserver's http client.

It's no longer necessary to pass the `use_ssl` option, as that will be
determined based on the URL passed in.