(PUP-10379) Replace environment loader with project loader in Bolt context
When loading modules as part of running Bolt we want to load the Bolt
project directory as a single module if `bolt_project` is configured.
The environment loader already behaves this way by creating a module
loader with the module name 'environment', then loading modules from
it's modulepath using a dependency loader. Because Bolt doesn't make use
of environments or have a concept of an environment, we're able to swap
in the hardcoded 'environment' module for the project directory to load
as a standalone module.

This is sufficient for loading the Bolt project, but because all modules
have the environment as a dependency when listing artifacts for the
project directory module it duplicates the entries for each module
instance. For example if you modulpath includes 33 modules, each task or
plan in the Bolt project "module" will be returned 33 times. Without
modifying how dependencies are loaded, the simplest way around this is
to deduplicate the list.