(PUP-10152) Prevent Pacman package provider doing partial upgrades
Using the `-y` switch in Pacman refreshes the catalog, but doesn't
upgrade existing packages. This causes a "partial upgrade".
From the Arch Linux wiki: [Partial upgrades are
unsupported](https://wiki.archlinux.org/index.php/System_maintenance#Partial_upgrades_are_unsupported)

This change removes the `-y` switch. It will be up to the user to
refresh the catalog and upgrade the system _explicitly_. The other
approach would be to have both the `-y` _and_ `-u` switches, but that
would mean that installing a package through Puppet would implicitly
upgrade the whole of the system - I don't think that would be desirable.