(PUP-10248) Windows pidlock can raise access denied
Prior to this commit, if puppet is running in the background as a
service, and the user runs `puppet agent -t` in the foreground, the
foreground process may not have permission to open the process token
for the background process running as `LocalSystem`, causing puppet to
raise with an ambiguous error message.

To fix this, we open the process token with the `SeDebugPrivilege`,
which allows access to any process.

If we still don't have access to open the process token, we assume
one of the following, and do not unlock the lockfile:

- Not running elevated (`Puppet.features.root?` is false), then
`with_privilege` just yields. So the call to `open_process` can fail
if the PID was started by a different user.

- Running elevated, but the `SeDebug` privilege isn't in our process
token, then trying to enable the privilege can fail (we can only
enable privileges that are disabled in our process token, but we can't
add privileges we weren't granted).

- Running elevated, but no permission to open the process token due to
it being privileged (csrss.exe)