(maint) Verify POST requests detect EOF
Prior to ruby 2.3, if puppet tried to reuse a persistent connection, but the
remote side had already closed it, then puppet would receive an EOFError.

Ruby 2.3.0 does a non-blocking read to detect if the connection is closed, and
automatically reconnects if needed. See ruby bug 11671 and commit [1].

Since puppet 6 requires 2.3.0 or greater, there isn't anything for puppet to do
for PUP-3238. However, this commit adjusts the persistent http test to POST
instead of GET. This is because ruby 2.1.5 and up will automatically retry
idempotent HTTP requests, like GET, which could mask the issue. Note when
running on ruby 2.5 and up, puppet disables the ruby's auto retry, so that the
decision to retry is not being made in multiple places. See PUP-3905 and commit
1e4fd336568afaf989e1e0b782629aa42903b8d5 for more context.

[1] https://github.com/ruby/ruby/commit/d249be623375e2725e1e3a979d7d47e45d356265