(PUP-10379) Load bolt project into environment module list
If the keyword `bolt_project` is defined, the Bolt project should be
loaded at the head of the `Puppet::Node::Environment#modules` list of
modules. Bolt sets a temporary environment when running which it uses to
load modules on it's modulepath. We want to maintain this behavior and
also load the Bolt project directory configure by the user as a
standalone module. This ensures that any modules with the same name as
the Bolt project aren't loaded, so that the Bolt project can safely
shadow existing modules only when specified in Bolt. Additionally this
ensures that only the Bolt project is loaded as a module, not any
sibling directories. The keyword should only ever be set in Bolt
contexts and should not impact Puppet environments.