(PUP-10385) Change filebucket terminus to use http client
Since we've created a new framework for handling HTTP requests, we need
to switch everything over to using that new framework. This commit
updates filebucket to do that.

We get to remove redundant logic around finding what server to send
requests to. However, because of the nature of filebucket, we need to
maintain some of the old strange behavior. We still need to pass a host
and post to Puppet::FileBucket::Dipper, so we have to look up the http
session and route to the puppet service both when we create the Dipper
and when we go to make the requests.